Task discription
Linear regression is a supervised learning algorithm that predicts a continuous output variable (in this case, house price) based on one or more input features (in this case, square footage, number of bedrooms, and number of bathrooms). The goal is to learn a linear relationship between the input features and the output variable.

Linear regression assumes that the relationship between the input features and the output variable is linear, and that the data is normally distributed. Additionally, it assumes that there is no multicollinearity between the input features (i.e., they are not highly correlated with each other).

Working:
-Data Preparation: The dataset is loaded and the input features (square footage, number of bedrooms, and number of bathrooms) and the output variable (house price) are selected.
-Data Splitting: The data is split into training and testing sets to evaluate the model's performance.
-Model Creation: A linear regression model is created and trained on the training data.
-Model Training: The model learns the linear relationship between the input features and the output variable by minimizing the mean squared error (MSE) between the predicted and actual values.
-Model Evaluation: The model is evaluated on the testing data using metrics such as MSE and R-squared.
-Prediction: The model is used to make predictions on new data.

Code Explanation:
-X and y are created by selecting the input features and the output variable from the dataset, respectively.
-train_test_split is used to split the data into training and testing sets.
-LinearRegression is used to create a linear regression model.
-fit is used to train the model on the training data.
-predict is used to make predictions on the testing data.
-mean_squared_error and r2_score are used to evaluate the model's performance.
-The model is used to make a prediction on new data using the predict method.